pyproject.toml
[tool.poetry]
name = "CarlAdam"
version = "0.1.0"
description = "Petri Net tools for Python"
authors = ["Matthew Scott <matt@11craft.com>"]

[tool.poetry.dependencies]
attrs = "^23.2.0"
multiset = "^3.0.2"
networkx = "^3.2.1"
numpy = "^1.26.4"
pyrsistent = "^0.20.0"
python = "^3.11"
python-decouple = "^3.8"

[tool.poetry.group.dev.dependencies]
Jinja2 = "^3.1.3"
Markdown = "^3.5.2"
coverage = "^7.4.3"
httpx = "^0.27.0"
pre-commit = "^3.6.2"
pytest = "^8.0.2"
pytest-cov = "^4.1.0"
types-Markdown = "^3.5.0"
watchdog = {extras = ["watchmedo"], version = "^4.0.0"}

[tool.poetry.group.simulator.dependencies]
django = "^5.0.3"

[build-system]
requires = ["poetry-core>=1.9.0"]
build-backend = "poetry.core.masonry.api"

[tool.coverage.run]
omit = [
  "*tests/*.py",
  # Non-core modules undergoing rapid change:
  "*diagram/graphviz.py",
  "*diagram/kroki.py",
  "*diagram/lab.py",
]
source = ["."]

[tool.coverage.html]
directory = "../public/static/htmlcov"

[tool.ruff]
line-length = 120
target-version = "py312"

[tool.ruff.lint.mccabe]
max-complexity = 10

[tool.ruff.lint.isort]
case-sensitive = true
lines-after-imports = 2
